fixup! evaluation: improve repetitions
Bench 3096883

+++ src/core/bit_board.h
@@ -134,6 +134,6 @@ struct BitBoard {
     std::optional<BoardPosition> enPessant;
 
     // amount of rounds that the game has been played
-    uint32_t fullMoves {};
-    uint32_t halfMoves {};
+    uint16_t fullMoves {};
+    uint16_t halfMoves {};
 };
